% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{continent_colours}
\alias{continent_colours}
\title{Colours for continents}
\format{
A string list with 13 entries.
\describe{
\item{Africa}{The colour for Africa, "orange".}
\item{Asia}{The colour for Asia, "gray".}
\item{Bunkers}{The colour for bunkers, "darkgray".}
\item{Europe}{The colour for Europe, "Europe".}
\item{MidEast, Middle East}{The colour for the Middle East, "cadetblue3".}
\item{NoAmr, North America}{The colour for North America, "red".}
\item{Oceania}{The colour for Oceania, "blue".}
\item{SoCeAmr, South and Central America}{The colour for South and Central America, "violet".}
\item{Africa}{The colour for World, "black".}
}
}
\usage{
continent_colours
}
\description{
A string list containing names of colours for continents.
These colours were used for Figure B1 of
P. E. Brockway, M. K. Heun, Z. Marshall, E. Aramendia,
P. Steenwyk, T. Relph, M. Widjanarko,
J. J. Kim, A. Sainju, and J. Irtube.
A country-level primary-final-useful (CL-PFU) energy and exergy
database: Overview of its construction and 1971–2020
world-level efficiency results.
Environmental Research: Energy, 1(025005):1–29, June 2024.
\url{https://doi.org/10.1088/2753-3751/ad4e39}
}
\examples{
continent_colours
}
\keyword{datasets}
