% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{continent_linetypes}
\alias{continent_linetypes}
\title{Continent linetypes}
\format{
A string list with 13 entries.
\describe{
\item{Africa}{The linetype for Africa, "solid".}
\item{Asia}{The linetype for Asia, "solid".}
\item{Bunkers}{The linetype for bunkers, "dashed".}
\item{Europe}{The linetype for Europe, "solid".}
\item{MidEast, Middle East}{The linetype for the Middle East, "solid".}
\item{NoAmr, North America}{The linetype for North America, "solid".}
\item{Oceania}{The linetype for Oceania, "solid".}
\item{SoCeAmr, South and Central America}{The linetype for South and Central America, "solid".}
\item{Africa}{The linetype for World, "solid".}
}
}
\usage{
continent_linetypes
}
\description{
A string list containing linetypes for continents.
These linetypes were used for Figure B1 of
P. E. Brockway, M. K. Heun, Z. Marshall, E. Aramendia,
P. Steenwyk, T. Relph, M. Widjanarko,
J. J. Kim, A. Sainju, and J. Irtube.
A country-level primary-final-useful (CL-PFU) energy and exergy
database: Overview of its construction and 1971–2020
world-level efficiency results.
Environmental Research: Energy, 1(025005):1–29, June 2024.
\url{https://doi.org/10.1088/2753-3751/ad4e39}
}
\examples{
continent_linetypes
}
\keyword{datasets}
