% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{continent_linewidths}
\alias{continent_linewidths}
\title{Continent linewidths}
\format{
A string list with 13 entries.
\describe{
\item{Africa}{The linewidth for Africa, 0.3.}
\item{Asia}{The linewidth for Asia, 0.3.}
\item{Bunkers}{The linewidth for bunkers, 0.2.}
\item{Europe}{The linewidth for Europe, 0.3.}
\item{MidEast, Middle East}{The linewidth for the Middle East, 0.3.}
\item{NoAmr, North America}{The linewidth for North America, 0.3.}
\item{Oceania}{The linewidth for Oceania, 0.3.}
\item{SoCeAmr, South and Central America}{The linewidth for South and Central America, 0.3.}
\item{Africa}{The linewidth for World, 0.3.}
}
}
\usage{
continent_linewidths
}
\description{
A string list containing linewieths for continents.
These linetypes were used for Figure B1 of
P. E. Brockway, M. K. Heun, Z. Marshall, E. Aramendia,
P. Steenwyk, T. Relph, M. Widjanarko,
J. J. Kim, A. Sainju, and J. Irtube.
A country-level primary-final-useful (CL-PFU) energy and exergy
database: Overview of its construction and 1971–2020
world-level efficiency results.
Environmental Research: Energy, 1(025005):1–29, June 2024.
https://doi.org/10.1088/2753-3751/ad4e39
}
\examples{
continent_linewidths
}
\keyword{datasets}
