% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primary_carrier_colours}
\alias{primary_carrier_colours}
\title{Primary-stage energy carrier colours}
\format{
A string list with 16 entries.
\describe{
\item{BM; Biomass}{The colour for Biomass.}
\item{CCP; Coal and coal products; Coal and coal\\nproducts}{The colour for Coal and coal products.}
\item{OOP; Oil and oil products; Oil and oil\\nproducts}{The colour for Oil and oil products.}
\item{NG; Natural gas}{The colour for Natural gas.}
\item{BF; Biofuels}{The colour for Biofuels.}
\item{Elect; Electricity}{The colour for Electricity.}
\item{Heat}{The colour for Heat.}
\item{Waste}{The colour for Waste.}
}
}
\usage{
primary_carrier_colours
}
\description{
A string list containing colours for primary-stage energy carriers.
These colours were used for Figure 5 of
P. E. Brockway, M. K. Heun, Z. Marshall, E. Aramendia,
P. Steenwyk, T. Relph, M. Widjanarko,
J. J. Kim, A. Sainju, and J. Irtube.
A country-level primary-final-useful (CL-PFU) energy and exergy
database: Overview of its construction and 1971–2020
world-level efficiency results.
Environmental Research: Energy, 1(025005):1–29, June 2024.
\url{https://doi.org/10.1088/2753-3751/ad4e39}
}
\examples{
primary_carrier_colours
}
\keyword{datasets}
