% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sector_colours}
\alias{sector_colours}
\title{Sector colours}
\format{
A string list with 21 entries.
\describe{
\item{AFF; Agriculture, forestry, and fishing; Agriculture, forestry,\\nand fishing}{The colour for the Agriculture, forestry, and fishing sector.}
\item{Res; Residential}{The colour for the Residential sector.}
\item{Ind; Industry}{The colour for the Industry sector.}
\item{CPS; Commercial and public services; Commercial and\\npublic services}{The colour for the Commercial and public services sector.}
\item{Trans; Transport}{The colour for the Transport sector.}
\item{NES}{The colour for the NES sector.}
\item{Final consumption not elsewhere specified; Final consumption not\\nelsewhere specified}{The colour for unspecified sectors.}
\item{NEU; Non-energy use}{The colour for the Non-energy use sector.}
\item{EIOU; Energy industry own use; Energy industry\\nown use}{The colour for the Energy industry own use sector.}
\item{World}{The colour for a World sector.}
}
}
\usage{
sector_colours
}
\description{
A string list containing colours for economic sectors.
These colours were used for Figure 6 of
P. E. Brockway, M. K. Heun, Z. Marshall, E. Aramendia,
P. Steenwyk, T. Relph, M. Widjanarko,
J. J. Kim, A. Sainju, and J. Irtube.
A country-level primary-final-useful (CL-PFU) energy and exergy
database: Overview of its construction and 1971–2020
world-level efficiency results.
Environmental Research: Energy, 1(025005):1–29, June 2024.
\url{https://doi.org/10.1088/2753-3751/ad4e39}
}
\examples{
sector_colours
}
\keyword{datasets}
