% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{useful_carrier_colours}
\alias{useful_carrier_colours}
\title{Useful-stage energy carrier colours}
\format{
A string list with 30 entries.
\describe{
\item{AnMech; Animal mechanical work; Animal mechanical\\nwork}{The colour for Animal mechanical work.}
\item{HuMech; Human mechanical work; Human mechanical\\nwork}{The colour for human mechanical work.}
\item{MechW; Mechanical work}{The colour for Mechanical work.}
\item{Prop; Propulsion}{The colour for propulsion.}
\item{HPF; High-pressure fluids; High-pressure\\nfluids}{The colour for High-pressure fluids.}
\item{LTH; Low-temperature heat; Low-temperature\\nheat}{The colour for Low-temperature heat.}
\item{MTH; Medium-temperature heat; Medium-temperature\\nheat}{The colour for medium-temperature heat.}
\item{HTH; High-temperature heat; High-temperature\\nheat}{The colour for High-temperature heat.}
\item{LTC; Low-temperature cooling; Low-temperature\\ncooling}{The colour for Low-temperature cooling.}
\item{IP; Information processing}{The colour for Information processing.}
\item{L; Light}{The colour for Light.}
}
}
\usage{
useful_carrier_colours
}
\description{
A string list containing colours for useful-stage energy carriers.
These colours were used for Figure 5 of
P. E. Brockway, M. K. Heun, Z. Marshall, E. Aramendia,
P. Steenwyk, T. Relph, M. Widjanarko,
J. J. Kim, A. Sainju, and J. Irtube.
A country-level primary-final-useful (CL-PFU) energy and exergy
database: Overview of its construction and 1971–2020
world-level efficiency results.
Environmental Research: Energy, 1(025005):1–29, June 2024.
\url{https://doi.org/10.1088/2753-3751/ad4e39}
}
\examples{
useful_carrier_colours
}
\keyword{datasets}
