% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{xy_theme}
\alias{xy_theme}
\title{Set a pleasing theme for graphs.}
\usage{
xy_theme(
  font_size = 12,
  font_family = "",
  tick_length = -0.3 * font_size,
  x_axis_labels_spacing = 0.2 * font_size,
  y_axis_labels_spacing = 0.2 * font_size,
  x_strip_labels_spacing = 0.2 * font_size,
  y_strip_labels_spacing = 0.2 * font_size,
  border_and_label_colour = "gray50"
)
}
\arguments{
\item{font_size}{The base font size in points for fonts and other graphical elements.
Default is \code{12}.}

\item{font_family}{The font family used in the graph.
Default is an empty string (\code{""}), a sans-serif font.}

\item{tick_length}{The tick length in points. Default is \code{-0.3*font_size}, which is -3.6 points.
Negative is inside the graph, where ticks belong!}

\item{x_axis_labels_spacing}{The distance between x axis labels and the x axis itself in points.
Default is \code{0.2*font_size}.}

\item{y_axis_labels_spacing}{The distance between y axis labels and the y axis itself in points.
Default is \code{0.2*font_size}.}

\item{x_strip_labels_spacing}{The distance between x strip text
(labels at the top of a faceted plot) and
the top border of the individual graphs
in points.
Default is \code{0.2*font_size} to leave space
for descenders.}

\item{y_strip_labels_spacing}{The distance between y strip text
(labels to the right of a faceted plot) and
the right border of the individual graphs
in points.
Default is \code{0.2*font_size} to leave space
for descenders.}

\item{border_and_label_colour}{The colour for borders and labels. Default is "gray50".}
}
\description{
This function provides a graphical style for \code{ggplot2} that includes
\itemize{
\item gray axes,
\item gray tic marks,
\item inward-pointing tic marks,
\item right-justified y-axis labels,
\item no grid lines, and
\item white background.
}
}
\examples{
library(ggplot2)
DF <- data.frame(x = seq(1, 10), y = seq(11,20))
p <- DF |>
  ggplot(mapping = aes(x = x, y = y)) +
  geom_line()
p
p + MKHthemes::xy_theme()
}
