% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{xy_theme}
\alias{xy_theme}
\title{Set the typical theme for graphs.}
\usage{
xy_theme(base_size = 12, base_family = "", base_theme = theme_bw,
  label_colour = "gray50")
}
\arguments{
\item{base_size}{the base font size for text elements. Default is \code{12}.}

\item{base_family}{the base font family. Default is an empty string (\code{""}).}

\item{base_theme}{the base theme for ggplot2 graphics. Default is \code{theme_by}.}

\item{label_colour}{the colour for labels. Default is "\code{gray50}".}
}
\description{
This function provides a graphical style that includes
\itemize{
  \item gray axes,
  \item gray tic marks,
  \item inward-pointing tic marks,
  \item no grid lines,
  \item and white plot background.#'
}
}
\examples{
library(ggplot2)
DF <- data.frame(x = seq(1, 10), y = seq(11,20))
ggplot(data = DF, mapping = aes(x = x, y = y)) +
  geom_point() +
  xy_theme()
}
